requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.0
wandb>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
PyYAML>=6.0
tqdm>=4.0.0
albumentations>=2.0.5